Add initial version of a round-robin memory access scheduler
The purpose is two-fold. First, this device won't be the only thing on the
memory bus, so the pipeline parts need to have a way to submit requests
and wait an arbitrary length of time for them to be fulfilled. Second, with
sprites, there won't be a fixed number of "threads" wanting to access RAM
anymore. Plus, there's an opportunity to save some effort if not all layers
are enabled.